+++ R/us-map.R
@@ -1,30 +1,28 @@
 #' Retrieve US map data
 #'
 #' @param regions The region breakdown for the map, can be one of
-#'   (\code{"states"}, \code{"state"}, \code{"counties"}, \code{"county"}).
-#'   The default is \code{"states"}.
-#' @param include The regions to include in the resulting map. If \code{regions} is
-#'  \code{"states"}/\code{"state"}, the value can be either a state name, abbreviation or FIPS code.
+#'   (`"states"`, `"state"`, `"counties"`, `"county"`).
+#'   The default is `"states"`.
+#' @param include The regions to include in the resulting map. If `regions` is
+#'  `"states"`/`"state"`, the value can be either a state name, abbreviation or FIPS code.
 #'  For counties, the FIPS must be provided as there can be multiple counties with the
 #'  same name. If states are provided in the county map, only counties in the included states
 #'  will be returned.
-#' @param exclude The regions to exclude in the resulting map. If \code{regions} is
-#'  \code{"states"}/\code{"state"}, the value can be either a state name, abbreviation or FIPS code.
+#' @param exclude The regions to exclude in the resulting map. If `regions` is
+#'  `"states"`/`"state"`, the value can be either a state name, abbreviation or FIPS code.
 #'  For counties, the FIPS must be provided as there can be multiple counties with the
-#'  same name. The regions listed in the \code{include} parameter are applied first and the
-#'  \code{exclude} regions are then removed from the resulting map. Any excluded regions
-#'  not present in the included regions will be ignored.
-#' @param as_sf Defunct, this parameter no longer has any effect and will be removed in
-#'  the future.
+#'  same name. The regions listed in the `include` parameter take precedence over
+#'  regions listed in `exclude`. If both parameters include the same region(s) they
+#'  will be included in the map.
 #' @param data_year The year for which to obtain map data.
-#' If the value is \code{NULL}, the most recent year's data is used. If the
-#' provided year is not found from the available map data sets, the next most
-#' recent year's data is used. This can be used if an older data set is being
-#' plotted on the US map so that the data matches the map more accurately.
-#' Therefore, the provided value should match the year of the plotted data set.
-#' The default is \code{NULL}, i.e. the most recent available year is used.
+#'  If the value is `NULL`, the most recent year's data is used. If the
+#'  provided year is not found from the available map data sets, the next most
+#'  recent year's data is used. This can be used if an older data set is being
+#'  plotted on the US map so that the data matches the map more accurately.
+#'  Therefore, the provided value should match the year of the plotted data set.
+#'  The default is `NULL`, i.e. the most recent available year is used.
 #'
-#' @return An `sf` data frame of US map coordinates divided by the desired \code{regions}.
+#' @return An `sf` data frame of US map coordinates divided by the desired `regions`.
 #'
 #' @examples
 #' str(us_map())
@@ -41,7 +39,6 @@ us_map <- function(
   regions = c("states", "state", "counties", "county"),
   include = c(),
   exclude = c(),
-  as_sf = TRUE,
   data_year = NULL
 ) {
   regions <- match.arg(regions)
@@ -53,18 +50,23 @@ us_map <- function(
   file_path <- system.file("extdata", map_year, file_name, package = "usmapdata")
   df <- sf::read_sf(file_path, as_tibble = FALSE)
 
-  if (length(include) > 0) {
-    df <- df[df$full %in% include |
-               df$abbr %in% include |
-               df$fips %in% include |
-               substr(df$fips, 1, 2) %in% include, ]
-  }
+  # remove excluded items that are in `include`
+  exclude <- setdiff(exclude, include)
 
+  # remove excludes
   if (length(exclude) > 0) {
     df <- df[!(df$full %in% exclude |
                  df$abbr %in% exclude |
                  df$fips %in% exclude |
                  substr(df$fips, 1, 2) %in% exclude), ]
+  }
+
+  # remove non-includes
+  if (length(include) > 0) {
+    df <- df[df$full %in% include |
+               df$abbr %in% include |
+               df$fips %in% include |
+               substr(df$fips, 1, 2) %in% include, ]
   }
 
   df[order(df$abbr), ]
@@ -75,12 +77,11 @@ us_map <- function(
 #' @inheritParams us_map
 #'
 #' @return An `sf` data frame of state or county centroid labels and positions
-#'   relative to the coordinates returned by the \code{us_map} function.
+#'   relative to the coordinates returned by the \link{us_map} function.
 #'
 #' @export
 centroid_labels <- function(
   regions = c("states", "state", "counties", "county"),
-  as_sf = TRUE,
   data_year = NULL
 ) {
   regions <- match.arg(regions)
@@ -109,8 +110,8 @@ available_map_years <- function() {
 
 #' Select appropriate map data year from available years
 #'
-#' @param data_year The year for which to obtain \code{usmap} data.
-#' If the value is \code{NULL}, the most recent year is returned. If the
+#' @param data_year The year for which to obtain \link{us_map} data.
+#' If the value is `NULL`, the most recent year is returned. If the
 #' provided year is not found from the available map data sets, the next most
 #' recent available year is returned.
 #'